feat(rpc-fastify): extract FastifyHttpAppServer module for general HttpApp-to-Fastify integration
