+++ packages/rpc-fastify/src/index.ts
@@ -1,4 +1,9 @@
 /**
  * @since 1.0.0
  */
+export * as FastifyHttpAppServer from "./FastifyHttpAppServer.js"
+
+/**
+ * @since 1.0.0
+ */
 export * as FastifyRpcServer from "./FastifyRpcServer.js"
